Several fixes.
Fixed issues due to missing configuration parameters. Now suitable
defaults are provided for non-mandatory parameters.
Modified driver instantiation to allow inheriting the requestGap and
requestTimeout values defined for a given gateway (only to connected
devices) and to use such values in case no register-specific value is
defined. Please notice that the requestTimeout can only be applied when
opening the serial connection, therefore it is currently useless at the
device level.