Fixed issue on scaling of register values. Bump version 1.2.2->1.2.3
During the read operation, registers are read and the associated scale
factor is applied (by multiplying the register value by the scale
factor). In this fix, the opposite path is also addressed using the
scale factor: given a value to write to a certain register, the value is
first scaled using the associated scale factor (dividing the value by
the scale factor) and then written to the register. It must be noticed
that since the scaling implies a conversion to floating point numbers,
the Math.round operator is applied to extract integer values, when
needed.